refactor: organize imports by moving types and modules
